% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartilage-package.R
\docType{package}
\name{cartilage-package}
\alias{cartilage}
\alias{cartilage-package}
\title{cartilage: Spatializing 'rayshader'}
\description{
Drive rayshader's image creation within a geospatial context. Input a spatial object
 (i.e. a raster) and build a record of the input object, the ray-shaded product along with a
 3D plot as a side-effect. This simplifies the integration of rayshader products with more 
 general tasks.
}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
