% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_rayshader.R
\name{rayshade_raster}
\alias{rayshade_raster}
\title{Restore rayshader to input array}
\usage{
rayshade_raster(shade, ras)
}
\arguments{
\item{shade}{rayshader array (built from 'ras')}

\item{ras}{input raster}
}
\description{
Rebuild rayshader RGB array to match its input.
}
\examples{
library(rayshader)
monterey_dem <- raster(system.file("extdata/monterey/monterey_dem.tif", package = "montereybayshader"))
r <- aggregate(monterey_dem, fact = 4)
r <- resample(r, raster(extent(r), nrows = 256, ncols = 206))
mdem <- t(as.matrix(r))
qm <- quadmesh::quadmesh(r)

## run any rayshader code here
sh <- mdem  \%>\%
sphere_shade(texture = "imhof1") \%>\%
add_shadow(ray_shade(mdem,zscale=200)) \%>\%
add_shadow(ambient_shade(mdem,zscale=200)) \%>\%
rayshade_raster(r)

library(rgl)
shade3d(qm, col = rep(rayshade_to_hex(sh), each = 4))
aspect3d(1, 1, .1)
}
