% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rayshade_to_hex}
\alias{rayshade_to_hex}
\title{Rayshader as RGB vector}
\usage{
rayshade_to_hex(x)
}
\arguments{
\item{x}{rayshader array}
}
\value{
vector of hex codes
}
\description{
Rayshader as RGB vector
}
